requirements.txt
asgiref==3.7.2
Django==3.2.19
django-cors-headers==4.0.0
djangorestframework==3.14.0
Pillow==9.0.0
gunicorn==21.2.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
